Change generateToken to add salt as optional on both providers
The token salt is an optional security technique that plugin users may
or may not implement, so both `LegacyStatelessTokenProvider` and
`JwtStatelessTokenProvider` have been modified to accept `salt` as
an **optional** parameter.

The methods themselves already supported `salt` being `null`, so nothing
more than the sign of the method have been changed, and tests added.